Move cookie code to @app/legal/cookies root
